+++ tests/test_resource.py
@@ -4,20 +4,20 @@ Test the behaviour of the new implementation while asserting the behaviour of th
 :copyright: 2021 Nándor Mátravölgyi
 :license: Apache2, see LICENSE for more details.
 """
+
 import asyncio
 
 import pytest
 
 import wait_for2
-from .common.constants import BUILTIN_PREFERS_CANCELLATION_OVER_RESULT
+from .common.constants import BUILTIN_WAIT_FOR_BEHAVIOUR, GT_PY312
 from .common.resource import ResourceWorkerWaitForTester
 
 
 @pytest.mark.asyncio
 async def test_resource_leakage_builtin():
-    if BUILTIN_PREFERS_CANCELLATION_OVER_RESULT:
-        with pytest.raises(AssertionError, match="resources were leaked"):
-            await ResourceWorkerWaitForTester(asyncio.wait_for).run()
+    if BUILTIN_WAIT_FOR_BEHAVIOUR["some timeout, cancel after "] == "cancelled bound":
+        await ResourceWorkerWaitForTester(asyncio.wait_for).run()
     else:
         with pytest.raises(AssertionError, match="wait_for within a task ignored the cancellation"):
             await ResourceWorkerWaitForTester(asyncio.wait_for).run()
@@ -32,12 +32,20 @@ async def test_resource_leakage_wf2_callback():
 
 @pytest.mark.asyncio
 async def test_resource_leakage_wf2_except():
-    # handle the cancellation-completion race condition as an exception
-    await ResourceWorkerWaitForTester(wait_for2.wait_for).run(use_special_raise=True)
+    if GT_PY312:
+        with pytest.raises(AssertionError, match="Special raise cleanup never called"):
+            await ResourceWorkerWaitForTester(wait_for2.wait_for).run(use_special_raise=True)
+    else:
+        # handle the cancellation-completion race condition as an exception
+        await ResourceWorkerWaitForTester(wait_for2.wait_for).run(use_special_raise=True)
 
 
 @pytest.mark.asyncio
 async def test_resource_leakage_wf2_no_handle():
-    # if we do not handle the race condition the alternate implementation is similar to the builtin
-    with pytest.raises(AssertionError, match="resources were leaked"):
+    if GT_PY312:
+        # since python 3.12 the builtin does not lose results (wait_for2 uses that automatically when no race_handler)
         await ResourceWorkerWaitForTester(wait_for2.wait_for).run()
+    else:
+        # if we do not handle the race condition the alternate implementation is similar to the builtin
+        with pytest.raises(AssertionError, match="resources were leaked"):
+            await ResourceWorkerWaitForTester(wait_for2.wait_for).run()